requirements.txt
Django==1.7.5
poplus-pci==0.1.6
geopy==1.9.1
django-environ==0.3.0